requirements.txt
# database and dbt requirements
dbt-duckdb

# additional packages for data generation
Faker
pandas
tqdm
